refactor: lines 캐싱전략 shortestPath로 이동
업데이트 시 캐싱 초기화 확인완료
