+++ src/main/java/wooteco/subway/path/domain/strategy/shortestpath/DijkstraShortestPathStrategy.java
@@ -1,7 +1,6 @@
 package wooteco.subway.path.domain.strategy.shortestpath;
 
 import org.jgrapht.GraphPath;
-import org.jgrapht.alg.interfaces.ShortestPathAlgorithm;
 import org.jgrapht.alg.shortestpath.DijkstraShortestPath;
 import org.jgrapht.graph.DefaultWeightedEdge;
 import wooteco.subway.line.domain.Line;
@@ -12,18 +11,7 @@ import java.util.List;
 
 public class DijkstraShortestPathStrategy extends ShortestPathStrategy {
     @Override
-    public List<Station> getVertexList(List<Line> lines, Station source, Station target) {
-        GraphPath<Station, DefaultWeightedEdge> graphPath = graphPath(lines, source, target);
-        return graphPath.getVertexList();
-    }
-
-    @Override
-    public int getWeight(List<Line> lines, Station source, Station target) {
-        GraphPath<Station, DefaultWeightedEdge> graphPath = graphPath(lines, source, target);
-        return (int) graphPath.getWeight();
-    }
-
-    private GraphPath<Station, DefaultWeightedEdge> graphPath(List<Line> lines, Station source, Station target) {
+    public GraphPath<Station, DefaultWeightedEdge> graphPath(List<Line> lines, Station source, Station target) {
         DijkstraShortestPath<Station, DefaultWeightedEdge> dijkstraShortestPath = new DijkstraShortestPath<>(new PathGraph(lines).graph());
         return dijkstraShortestPath.getPath(source, target);
     }
